Fix empty entries in mime.types.nginx
There are a few lines in mime.types that have no file extension, but
have trailing whitespace at the end.  The intent of the final regex in
the `sed` command in generate-nginx-mimetypes.sh is clearly to only
match lines that have file extensions; but for the described lines has
a false-positive, mistaking the inserted trailing-";" for a file
extension.

Fix this by changing the file-extension-detecting part of the regex
from

    [^[:space:]]

to

    [^[:space:];]